% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify body masses into size classes}
\usage{
classify(sizes, method = "Brain")
}
\arguments{
\item{sizes}{vector of body sizes to classify, units are assumed to be kg}

\item{method}{method of classification, default is Brain}
}
\description{
Classify body masses into size classes
}
\examples{
x <- runif(10, min=10, max=1500)
x[11] <- NA
x[12] <- 80
classify(x)
}
